feat(evolve): stop a fit and start it again as though it never stopped
The benchmark is 133 laws at ten seeds and eight hours a fit: 1,330 runs, a
month of GPU, on a laptop with other work to do. A run that cannot be stopped is
a run that cannot be finished — and one whose result depends on WHEN it was
stopped is not reproducible, so nothing built on it can be published.

Engine::fit now writes its whole state to the rotating slots every
EVOLVE_CHECKPOINT_EVERY seconds and at the end, and takes a checkpoint back on
its first beat: the population, the generation the generator is keyed on, the
cohort ages, HFF's frozen column maxima, the hall of fame, the counters, and the
seconds already spent. The budget counts the WHOLE fit's time, not this
process's, or a stopped run would be given more time than an uninterrupted one
and 2T would not equal T + T.

Seconds, not generations: a generation's cost changes with the population and
the benchmark budgets a fit in time.

PROVEN, on the law and on a toy. a_stopped_fit_equals_an_uninterrupted_one holds
24 generations against 12 + stop + resume + 12 and compares the MODEL, and it
was checked by sabotage — resuming one generation early makes it fail with 'the
resumed fit found a different model'. End to end at 40 s on strogatz_bacres1:
straight through gives 833 generations, early_stop, 1-R2 9.282e-11; stopped at
15 s and resumed gives 833, early_stop, 9.282e-11. Identical.

506 tests pass with --features gpu, clippy clean.

+++ src/evolve/checkpoint.rs
@@ -92,7 +92,7 @@ pub struct SavedHof {
     pub fitness: f64,
     pub one_minus_r2: [f64; 3],
     pub t_depth: u32,
-    pub linker: u32,
+    pub linker: usize,
     pub wrapper: usize,
     pub a: f64,
     pub b: f64,
